Fix NuGet errors when consumed by apps (#297)
This PR fixes our consumption of the `Microsoft.Windows.CppWinRT` package reference to use `Update` instead of `Include` (so at best it "updates" which version to use and doesn't conflict with what RNW wants) and removes the "duplicate reference" warning our customers see.

I'm 99% we don't need the entry at all in newer RNW versions but leaving it for now to deal with older customers.

This PR also removes the NuGet `package.lock.json` file from the published npm package, as it takes the decision to use/respect lock files away from the customer.